Debug save.
While adding a raycast method to the weapon manager I managed to get it to work and spawn an instance of a bullet hole decal, however I noticed the camera was not clamping properly after fixing that my raycast broke again. saving here so that I can branch off and mess with a few varaibles that might break a few things.